+++ examples/example_intersight_api_params.json
@@ -1,0 +1,5 @@
+{
+    "api_base_uri": "https://intersight.com/api/v1",
+    "api_key_id": "<api key id from Intersight portal>",
+    "api_private_key_file": "<directory path to>SecretKey.txt"
+}
