--- examples/example_settings.json
@@ -1,5 +1,0 @@
-{
-    "api_host": "https://intersight.com/api/v1",
-    "api_key_id": "<api key id>",
-    "api_private_key": "./SecretKey.txt"
-}
