+++ examples/get_data_3.pyc
